pyproject.toml
[project]
name = "fastapi-dev"
version = "0.1.0"
description = "FastAPI backend with Terraform and AWS deployment workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<4.1",
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,c,pool]>=3.2.7",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy[asyncio]>=2.0.40",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
